pyproject.toml
[project]
name = "llm-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "mpi4py>=4.1.1",
    "nvshmem4py-cu12>=0.3.0",
    "ray[default,serve]>=2.55.1",
    "safetensors>=0.7.0",
    "vllm>=0.19.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm_rl"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
typeCheckingMode = "off"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
